Extract IntervalSlider control to separate file
